refactor: Fixing auction data types and DTO field naming
Refines auction data structures for improved efficiency and consistency.

Updates `quantity` fields from `Long` to `Int` and `petLevel` fields from `Int` to `Byte` across DBOs, DTOs, and mappers. These changes reduce memory and storage footprint, especially beneficial for handling large volumes of auction data.

Also, refactors `AuctionItemDTO` property names from snake_case to camelCase (e.g., `bonus_lists` to `bonusLists`), utilizing `@JsonProperty` to maintain compatibility with external data sources. This improves adherence to Kotlin's naming conventions.

Removes the redundant `@MapsId` annotation from the `Auction` entity for minor cleanup.

+++ backend/src/main/kotlin/net/jonasmf/auctionengine/repository/rds/AuctionStatsHourlyJDBCRepository.kt
@@ -14,7 +14,7 @@ data class HourlyStatsUpsertRow(
     val modifierKey: String,
     val bonusKey: String,
     val price: Long?,
-    val quantity: Long?,
+    val quantity: Int?,
 )
 
 @Repository
